Python wrapper for CKKSTensor + some refactor
